enforce/inform mode can be set for each constraint (#351)
* move constraint enforce setting into constraint parameter


* update operator-sdk version
