+++ integrity-shield-operator/bundle/manifests/apis.integrityshield.io_integrityshields.yaml
@@ -1066,13 +1066,6 @@ spec:
                         type: array
                     type: object
                 type: object
-              constraintConfig:
-                type: string
-              constraintConfigKey:
-                description: constraint config
-                type: string
-              constraintConfigName:
-                type: string
               labels:
                 additionalProperties:
                   type: string
