+++ integrity-shield-operator/config/samples/apis_v1_integrityshield.yaml
@@ -24,8 +24,8 @@ spec:
   shieldApi:
     name: integrity-shield-api
     imagePullPolicy: Always
-    # image: quay.io/open-cluster-management/integrity-shield-api:0.2.5
-    image: localhost:5000/integrity-shield-api:0.2.5
+    image: quay.io/open-cluster-management/integrity-shield-api:0.2.5
+    # image: localhost:5000/integrity-shield-api:0.2.5
     port: 8080
     selector:     
       app: integrity-shield-api
@@ -36,6 +36,9 @@ spec:
   requestHandlerConfigKey: config.yaml
   requestHandlerConfigName: request-handler-config
   requestHandlerConfig: |
+    action:
+      inform: true
+      enforce: false
     sideEffect: 
       createDenyEvent: true
     log:
@@ -406,8 +409,8 @@ spec:
     enabled: true
     name: integrity-shield-observer
     imagePullPolicy: Always
-    # image: quay.io/open-cluster-management/integrity-shield-observer:0.2.5
-    image: localhost:5000/integrity-shield-observer:0.2.5
+    image: quay.io/open-cluster-management/integrity-shield-observer:0.2.5
+    # image: localhost:5000/integrity-shield-observer:0.2.5
     selector:     
       app: integrity-shield-observer
     logLevel: info
@@ -415,11 +418,4 @@ spec:
     exportDetailResult: true
     resultDetailConfigName: verify-resource-result
     resultDetailConfigKey: "config.yaml"
-  constraintConfigKey: "config.yaml"
-  constraintConfigName:  constraint-config
-  constraintConfig: |
-    constraints: 
-    - constraintName: "*"
-      action:
-        inform: true
-        enforce: false
+
