+++ integrity-shield-operator/config/samples/apis_v1_integrityshield_ac.yaml
@@ -19,6 +19,9 @@ spec:
   requestHandlerConfigKey: config.yaml
   requestHandlerConfigName: request-handler-config
   requestHandlerConfig: |
+    action:
+      inform: true
+      enforce: false
     sideEffect: 
       createDenyEvent: true
     log:
@@ -357,11 +360,4 @@ spec:
     exportDetailResult: true
     resultDetailConfigName: verify-resource-result
     resultDetailConfigKey: "config.yaml"
-  constraintConfigKey: "config.yaml"
-  constraintConfigName:  constraint-config
-  constraintConfig: |
-    constraints: 
-    - constraintName: "*"
-      action:
-        inform: true
-        enforce: false
+
